Created a DebianParser class, removed Entrypoint abstraction
Created a debian module with a DebianParser class. This class would be used as a parser for retrieving patches from Debian. Removed the Entrypoint abstraction entirely but left the Provider abstraction intact. This is because the default parse method in DefaultSpider only needs the xpath(s) for the entrypoint, which can be achieved from a mapping function.